write/elf: overflow checks (#918)
The general strategy is to check the maximum for a value instead
of checking every conversion.

Where possible, change functions to use narrower types when we
require the caller to have already validated. The exception is
for values that have different widths for 32-bit and 64-bit ELF.

The status is:

- Encoder types are narrowed where possible, and documented for others
- Writer checks file size, and has narrowed types for caller provided
  counts, but doesn't check accumulators for its own counts
- Builder checks file size and all table counts

+++ crates/examples/src/elfcopy.rs
@@ -393,7 +393,7 @@ pub fn elfcopy<Elf: FileHeader<Endian = Endianness>>(
     } else {
         // We don't support moving program headers.
         assert_eq!(in_elf.e_phoff(endian).into(), writer.reserved_len());
-        writer.reserve_program_headers(in_segments.len());
+        writer.reserve_program_headers(in_segments.len() as u32);
 
         // Reserve alloc sections at original offsets.
         alloc_sections = in_sections
